working on modularizing the code
